refactor(constants): Refactor constants and complexity options into a dedicated constants file
